feat: implementation of new batch_solr_request
