+++ impc_api_helper/impc_api_helper/iterator_solr_request.py
@@ -1,145 +1,160 @@
-import csv
+from IPython.display import display
 import json
-
+import pandas as pd
 import requests
+from tqdm import tqdm
+from .solr_request import solr_request
+from pathlib import Path
 
+def batch_solr_request(core, params, download=False, batch_size=5000, path_to_download='./'):
+    # Set params for batch request
+    params["start"] = 0  # Start at the first result
+    params["rows"] = batch_size  # Fetch results in chunks of 5000
 
-# Helper function to fetch results. This function is used by the 'iterator_solr_request' function.
-def entity_iterator(base_url, params):
-    """Generator function to fetch results from the SOLR server in chunks using pagination
+    # If user did not specify format, defaults to json.
+    if params.get("wt") is None:
+        params["wt"] = "json"
 
-    Args:
-        base_url (str): The base URL of the Solr server to fetch documents from.
-        params (dict): A dictionary of parameters to include in the GET request. Must include
-                       'start' and 'rows' keys, which represent the index of the first document
-                       to fetch and the number of documents to fetch per request, respectively.
+    # Check if it's multiple request
+    if params.get("field_list") is not None:
+        # Extract entities_list and entity_type from params
+        field_list = params.pop("field_list")
+        field_type = params.pop("field_type")
 
-    Yields:
-        dict: The next document in the response from the Solr server.
-    """
-    # Initialise variable to check the first request
-    first_request = True
-
-    # Call the API in chunks and yield the documents in each chunk
-    while True:
-        response = requests.get(base_url, params=params)
-        data = response.json()
-        docs = data["response"]["docs"]
-
-        # Print the first request only
-        if first_request:
-            print(f"Your first request: {response.url}")
-            first_request = False
-
-        # Yield the documents in the current chunk
-        for doc in docs:
-            yield doc
-
-        # Check if there are more results to fetch
-        start = params["start"] + params["rows"]
-        num_found = data["response"]["numFound"]
-        if start >= num_found:
-            break
-
-        # Update the start parameter for the next request
-        params["start"] = start
-
-    # Print last request and total number of documents retrieved
-    print(f"Your last request: {response.url}")
-    print(f'Number of found documents: {data["response"]["numFound"]}\n')
-
-
-# Function to iterate over field list and write results to a file.
-def iterator_solr_request(
-    core, params, filename="iteration_solr_request", format="json"
-):
-    """Function to fetch results in batches from the Solr API and write them to a file.
-    Defaults to fetching 5000 rows at a time.
-    Avoids cluttering local memory, ideal for large requests.
-    
-    Args:
-        core (str): The name of the Solr core to fetch results from.
-        params (dict): A dictionary of parameters to use in the filter query. Must include
-                       'field_list' and 'field_type' keys, which represent the list of field items (i.e., list of MGI model identifiers)
-                        to fetch and the type of the field (i.e., model_id) to filter on, respectively.
-        filename (str): The name of the file/path to write the results to. Defaults to 'iteration_solr_request'.
-        format (str): The format of the output file. Can be 'csv' or 'json'. Defaults to 'json'.
-
-    Returns: None
-
-    Example use case:
-        # List of model IDs.
-        models = ['MGI:3587188', 'MGI:3587185', 'MGI:3605874', 'MGI:2668213']
-
-        # Call iterator function
-        iterator_solr_request(
-            core='phenodigm',
-            params = {
-                'q': 'type:disease_model_summary',
-                'fl': 'model_id,marker_id,disease_id',
-                'field_list': models,
-                'field_type': 'model_id'
-            },
-            filename='model_ids',
-            format='csv'
+        # Construct the filter query with grouped model IDs
+        fq = "{}:({})".format(
+            field_type, " OR ".join(["{}".format(id) for id in field_list])
         )
-    """
+        # Show users the field and field values they passed to the function
+        print("Queried field:", fq)
+        # Set internal params the users should not change
+        params["fq"] = fq
 
-    # Validate format
-    if format not in ["json", "csv"]:
-        raise ValueError("Invalid format. Please use 'json' or 'csv'")
+    # Determine the total number of rows. Note that we do not request any data (rows = 0).
+    num_results, _ = solr_request(
+        core=core, params={**params, "start": 0, "rows": 0, "wt": "json"}, silent=True
+    )
+    print(f"Number of found documents: {num_results}")
 
-    # Base URL
+    # Download/display only logic
+    # If user decides to download, a generator is used to fetch data in batches without storing results in memory.
+    if download:
+        # Implement loop behaviour
+        filename = Path(path_to_download) / f"{core}.{params['wt']}"
+        gen = _batch_solr_generator(core, params, num_results)
+        _solr_downloader(params, filename, gen)
+        print(f"File saved as: {filename}")
+        return None
+
+    # If the number of results is small enough and download is off, it's okay to show as df
+    if num_results < 1000000 and not download:
+        return _batch_to_df(core, params, num_results)
+
+    # If it's too big, warn the user and ask if they want to proceed.
+    else:
+        print(
+            "Your request might exceed the available memory. We suggest setting 'download=True' and reading the file in batches"
+        )
+        prompt = input(
+            "Do you wish to proceed anyway? press ('y' or enter to proceed) / type('n' or 'exit' to cancel)"
+        )
+        match prompt:
+            case "n" | "exit":
+                print("Exiting gracefully")
+                exit()
+            case "y" | '':
+                print("Fetching data...")
+                return _batch_to_df(core, params, num_results)
+
+
+
+
+# Helper batch_to_df
+def _batch_to_df(core, params, num_results):
+    start = params["start"]
+    batch_size = params["rows"]
+    chunks = []
+    # If the 'wt' param was changed by error, we set it to 'json'
+    params["wt"] = "json"
+
+    # Request chunks until we have complete data.
+    with tqdm(total=num_results) as pbar:
+        while start < num_results:
+            # Update progress bar with the number of rows requested.
+            pbar.update(batch_size)
+
+            # Request chunk. We don't need num_results anymore because it does not change.
+            _, df_chunk = solr_request(
+                core=core,
+                params={**params, "start": start, "rows": batch_size},
+                silent=True,
+            )
+
+            # Record chunk.
+            chunks.append(df_chunk)
+            # Increment start.
+            start += batch_size
+        # Prepare final dataframe.
+        return pd.concat(chunks, ignore_index=True)
+
+
+def _batch_solr_generator(core, params, num_results):
     base_url = "https://www.ebi.ac.uk/mi/impc/solr/"
     solr_url = base_url + core + "/select"
+    start = params["start"]
+    batch_size = params["rows"]
 
-    # Extract entities_list and entity_type from params
-    field_list = params.pop("field_list")
-    field_type = params.pop("field_type")
+    with tqdm(total=num_results) as pbar:
+        while start <= num_results:
+            params["start"] = start
+            response = requests.get(solr_url, params=params, timeout=10)
 
-    # Construct the filter query with grouped model IDs
-    fq = "{}:({})".format(
-        field_type, " OR ".join(['"{}"'.format(id) for id in field_list])
-    )
+            if response.status_code == 200:
+                if params.get("wt") == "json":
+                    data = response.json()["response"]["docs"]
+                else:
+                    data = response.text
+                
+                # Update and refresh the progress bar after getting the data
+                pbar.update(batch_size)
+                pbar.refresh()
+                yield data
+                
+            else:
+                raise Exception(f"Request failed. Status code: {response.status_code}")
 
-    # Show users the field and field values they passed to the function
-    print("Queried field:", fq)
-    # Set internal params the users should not change
-    params["fq"] = fq
-    params["wt"] = "json"
-    params["start"] = 0  # Start at the first result
-    params["rows"] = 5000  # Fetch results in chunks of 5000
+            # pbar.update(batch_size)
+            start += batch_size
+        print(f"Your request URL after the last call:{response.url}")
 
-    try:
-        # Fetch results using a generator function
-        results_generator = entity_iterator(solr_url, params)
-    except Exception as e:
-        raise Exception("An error occurred while downloading the data: " + str(e))
 
-    # Append extension to the filename
-    filename = f"{filename}.{format}"
+# File writer
+def _solr_downloader(params, filename, solr_generator):
+    with open(filename, "w", encoding="UTF-8") as f:
+        
+        if params.get("wt") == "json":
+            f.write("[\n")
+            first_chunk = True
 
-    try:
-        # Open the file in write mode
-        with open(filename, "w", newline="") as f:
-            if format == "csv":
-                writer = None
-                for item in results_generator:
-                    # Initialize the CSV writer with the keys of the first item as the field names
-                    if writer is None:
-                        writer = csv.DictWriter(f, fieldnames=item.keys())
-                        writer.writeheader()
-                    # Write the item to the CSV file
-                    writer.writerow(item)
-                    # Write to json without loading to memory.
-            elif format == "json":
-                f.write("[")
-                for i, item in enumerate(results_generator):
-                    if i != 0:
-                        f.write(",")
-                    json.dump(item, f)
-                f.write("]")
-    except Exception as e:
-        raise Exception("An error occurred while writing the file: " + str(e))
+            for chunk in solr_generator:
+                # print('CHUNK',chunk,'\n')
+                for item in chunk:
+                    if not first_chunk:
+                        f.write(",\n")
+                    # print('ITEM',item)
+                    json.dump(item, f, ensure_ascii=False)
+                    first_chunk = False
+            f.write("\n]\n")
 
-    print(f"File {filename} was created.")
+        elif params.get("wt") == "csv":
+            first_chunk = True
+            for chunk in solr_generator:
+                lines = chunk.splitlines()
+                if first_chunk:
+                    # Write all lines in the first chunk
+                    f.write(chunk)
+                    first_chunk = False
+                else:
+                    # Skip the first line (header) in subsequent chunks
+                    f.write('\n' + '\n'.join(lines[1:]) + '\n')
+
